[wip] bid = max_priority_fee
additionally, use Provider batch cals for parallel queries

+++ examples/optimism-tx-auction/contract/src/Auction.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.13;
 import {requireTimeBefore} from "pod-sdk/Time.sol";
 
 contract Auction {
-    event BidSubmitted(uint256 indexed deadline, uint256 bid, bytes data);
+    event BidSubmitted(uint256 indexed deadline, uint128 bid, bytes data);
 
-    function submitBid(uint256 deadline, uint256 bid, bytes calldata data) external {
+    function submitBid(uint256 deadline, uint128 bid, bytes calldata data) external {
         require(bid > 0, "Bid amount must be greater than zero");
         requireTimeBefore(deadline, "Auction deadline passed");
